glamor: split gbm-related functions from glamor_egl
Moved the gbm ifdef jungle out of glamor_egl
